0.2

+++ src/main/resources/db/changelog/v-1.0/20-create-items-table.sql
@@ -2,7 +2,7 @@ CREATE TABLE items(
     id serial PRIMARY KEY ,
     name VARCHAR (255),
     description TEXT,
-    price DOUBLE ,
+    price NUMERIC,
     department_id INTEGER REFERENCES departments(id) ON DELETE RESTRICT
 );
